made changes and resolved conflicts
